End point of 2/24, Should be good
As requested by Isaac, we finished touching up some code, and it looks decent. We need engineering to specify things such as the Right Motor Port, Left Motor Port, each motor's power attributes (positive or negative power), and Encoder Port 1. Ian (Haifeng) was supposedly told that engineering needed a "shortcut" for the low goal in the comp., so we need angle values in terms of buttons and the preset arm positioning (see code). Other than that, that's all we have for tonight.